i32 is now isize

+++ src/model/gamedata/point_map.rs
@@ -5,11 +5,11 @@ use crate::model::card::card::{Card};
 #[derive(Debug, Clone)]
 pub struct PointMap {
     pub name: String,
-    pub entries: HashMap<String, Vec<i32>>,
+    pub entries: HashMap<String, Vec<isize>>,
 }
 
 impl PointMap {
-    pub fn get_card_value_ref(&self, card: &Card) -> Option<Vec<i32>> {
+    pub fn get_card_value_ref(&self, card: &Card) -> Option<Vec<isize>> {
         for (_, value) in &card.attributes {
             if let Some(score) = self.entries.get(value) {                
                 return Some(score.clone());
@@ -19,7 +19,7 @@ impl PointMap {
         None
     }
 
-    pub fn get_card_value(&self, card: Card) -> Option<Vec<i32>> {
+    pub fn get_card_value(&self, card: Card) -> Option<Vec<isize>> {
         for (_, value) in card.attributes {
             if let Some(score) = self.entries.get(&value) {
                 return Some(score.clone());
